requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
langchain-openai
langchain-gigachat
python-dotenv
pyjwt
requests
langchain